+++ starter-code/package.json
@@ -3,24 +3,29 @@
   "version": "0.0.0",
   "private": true,
   "scripts": {
-    "start": "node ./bin/www"
+    "start": "node ./bin/www",
+    "mon": "nodemon ./bin/www"
   },
   "dependencies": {
     "bcrypt": "^1.0.2",
-    "body-parser": "~1.15.2",
+    "body-parser": "^1.15.2",
     "connect-ensure-login": "^0.1.1",
     "connect-flash": "^0.1.1",
-    "cookie-parser": "~1.4.3",
+    "cookie-parser": "^1.4.3",
     "debug": "~2.2.0",
+    "dotenv": "^4.0.0",
     "ejs": "~2.5.2",
     "express": "~4.14.0",
     "express-ejs-layouts": "^2.2.0",
-    "express-session": "^1.15.1",
-    "mongoose": "^4.8.2",
-    "morgan": "~1.7.0",
+    "express-session": "^1.15.2",
+    "mongoose": "^4.9.9",
+    "morgan": "^1.7.0",
     "multer": "^1.3.0",
     "passport": "^0.3.2",
     "passport-local": "^1.0.0",
     "serve-favicon": "~2.3.0"
+  },
+  "devDependencies": {
+    "nodemon": "^1.11.0"
   }
 }
