+++ starter-code/routes/authentication.js
@@ -1,10 +1,21 @@
 const express    = require('express');
+const bcrypt     = require('bcrypt');
 const passport   = require('passport');
-const router     = express.Router();
 const { ensureLoggedIn, ensureLoggedOut } = require('connect-ensure-login');
+const multer     = require('multer');
+const path       = require('path');
+
+const User = require('../models/user.js');
+
+const router     = express.Router();
+
 
 router.get('/login', ensureLoggedOut(), (req, res) => {
     res.render('authentication/login', { message: req.flash('error')});
+});
+
+const myUploader = multer({
+  dest: path.join(__dirname, '../public/uploads')
 });
 
 router.post('/login', ensureLoggedOut(), passport.authenticate('local-login', {
@@ -13,8 +24,11 @@ router.post('/login', ensureLoggedOut(), passport.authenticate('local-login', {
   failureFlash : true
 }));
 
-router.get('/signup', ensureLoggedOut(), (req, res) => {
-    res.render('authentication/signup', { message: req.flash('error')});
+router.get('/signup',
+ ensureLoggedOut(),
+ (req, res) => {
+    res.render('authentication/signup',
+    { message: req.flash('error')});
 });
 
 router.post('/signup', ensureLoggedOut(), passport.authenticate('local-signup', {
