+++ starter-code/views/authentication/login.ejs
@@ -1,7 +1,13 @@
+<h2> Log In </h2>
+
+<% if (errorMessage.length > 0) { %>
+  <div class="error"> <%= errorMessage %> </div>
+<% } %>
+
 <div class="container">
   <div class="row">
     <div class="col-sm-6 centered-form">
-      <form action="/login" method="post">
+      <form method="post" action="/login">
         <div class="form-group">
           <label for="username">Username</label>
           <input type="text" name="username" class="form-control">
