+++ starter-code/views/authentication/signup.ejs
@@ -1,22 +1,27 @@
 <div class="container">
   <div class="row">
     <%= message %>
-    <div class="col-sm-6 centered-form">
-      <form action="/signup" method="post">
-        <div class="form-group">
-          <label for="username">Username</label>
-          <input type="text" name="username" class="form-control">
-        </div>
-        <div class="form-group">
-          <label for="email">Email</label>
-          <input type="text" name="email" class="form-control">
-        </div>
-        <div class="form-group">
-          <label for="password">Password</label>
-          <input type="password" name="password" class="form-control">
-        </div>
-        <button type="submit">Signup</button>
-      </form>
-    </div>
+      <div class="col-sm-6 centered-form">
+        <form action="/signup" method="post">
+          <div class="form-group">
+            <label for="username">Username</label>
+            <input type="text" name="username" class="form-control">
+          </div>
+          <div class="form-group">
+            <label for="email">Email</label>
+            <input type="text" name="email" class="form-control">
+          </div>
+          <div class="form-group">
+            <label for="password">Password</label>
+            <input type="password" name="password" class="form-control">
+          </div>
+          <div class="form-group">
+            <label for="photo-input"> Upload Photo</label>
+            <input id="photo-input" type="file" name="photo">
+          </div>
+          <br>
+          <button type="submit">Signup</button>
+        </form>
+      </div>
   </div>
 </div>
